making errors more clear
